# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * 晟風工程 ERP API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Customer {
  id: number;
  name: string;
  phone: string;
  address: string;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  discountScheme?: string | null;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface CustomerInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  phone: string;
  /** @minLength 1 */
  address: string;
  email?: string;
  discountScheme?: string;
  notes?: string;
}

export interface CustomerUpdate {
  /** @minLength 1 */
  name?: string;
  phone?: string;
  address?: string;
  email?: string;
  discountScheme?: string;
  notes?: string;
}

export interface AcUnit {
  id: number;
  customerId: number;
  brand: string;
  model: string;
  /** @nullable */
  serialNumber?: string | null;
  /** @nullable */
  purchaseDate?: string | null;
  /** @nullable */
  installationDate?: string | null;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface AcUnitInput {
  /** @minLength 1 */
  brand: string;
  /** @minLength 1 */
  model: string;
  serialNumber?: string;
  purchaseDate?: string;
  installationDate?: string;
  notes?: string;
}

export interface AcUnitUpdate {
  brand?: string;
  model?: string;
  serialNumber?: string;
  purchaseDate?: string;
  installationDate?: string;
  notes?: string;
}

export interface Quote {
  id: number;
  customerId: number;
  /** @nullable */
  customerName?: string | null;
  title: string;
  /** @nullable */
  description?: string | null;
  amount: number;
  /** @nullable */
  discountAmount?: number | null;
  /** @nullable */
  finalAmount?: number | null;
  status: string;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface QuoteInput {
  customerId: number;
  /** @minLength 1 */
  title: string;
  description?: string;
  amount: number;
  discountAmount?: number;
  finalAmount?: number;
  status: string;
  notes?: string;
}

export interface QuoteUpdate {
  title?: string;
  description?: string;
  amount?: number;
  discountAmount?: number;
  finalAmount?: number;
  status?: string;
  notes?: string;
}

export interface WorkOrder {
  id: number;
  customerId: number;
  /** @nullable */
  customerName?: string | null;
  /** @nullable */
  quoteId?: number | null;
  title: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  assignedTo?: string | null;
  /** @nullable */
  scheduledDate?: string | null;
  /** @nullable */
  completedDate?: string | null;
  status: string;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface WorkOrderInput {
  customerId: number;
  quoteId?: number;
  /** @minLength 1 */
  title: string;
  description?: string;
  assignedTo?: string;
  scheduledDate?: string;
  completedDate?: string;
  status: string;
  notes?: string;
}

export interface WorkOrderUpdate {
  title?: string;
  description?: string;
  assignedTo?: string;
  scheduledDate?: string;
  completedDate?: string;
  status?: string;
  notes?: string;
}

export interface Progress {
  id: number;
  workOrderId: number;
  description: string;
  /** @nullable */
  recordedBy?: string | null;
  createdAt: string;
}

export interface ProgressInput {
  /** @minLength 1 */
  description: string;
  recordedBy?: string;
}

export interface Payment {
  id: number;
  customerId: number;
  /** @nullable */
  customerName?: string | null;
  /** @nullable */
  quoteId?: number | null;
  /** @nullable */
  workOrderId?: number | null;
  amount: number;
  paymentDate: string;
  /** @nullable */
  paymentMethod?: string | null;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}

export interface PaymentInput {
  customerId: number;
  quoteId?: number;
  workOrderId?: number;
  amount: number;
  paymentDate: string;
  paymentMethod?: string;
  notes?: string;
}

export interface PaymentUpdate {
  amount?: number;
  paymentDate?: string;
  paymentMethod?: string;
  notes?: string;
}

export interface Warranty {
  id: number;
  customerId: number;
  /** @nullable */
  customerName?: string | null;
  /** @nullable */
  acUnitId?: number | null;
  startDate: string;
  endDate: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}

export interface WarrantyInput {
  customerId: number;
  acUnitId?: number;
  startDate: string;
  endDate: string;
  description?: string;
  notes?: string;
}

export interface WarrantyUpdate {
  startDate?: string;
  endDate?: string;
  description?: string;
  notes?: string;
}

export interface MaintenanceReminder {
  id: number;
  customerId: number;
  /** @nullable */
  customerName?: string | null;
  /** @nullable */
  acUnitId?: number | null;
  reminderDate: string;
  description: string;
  status: string;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}

export interface MaintenanceReminderInput {
  customerId: number;
  acUnitId?: number;
  reminderDate: string;
  /** @minLength 1 */
  description: string;
  status: string;
  notes?: string;
}

export interface MaintenanceReminderUpdate {
  reminderDate?: string;
  description?: string;
  status?: string;
  notes?: string;
}

export interface DashboardSummary {
  totalCustomers: number;
  totalQuotes: number;
  totalWorkOrders: number;
  pendingWorkOrders: number;
  inProgressWorkOrders?: number;
  completedWorkOrders?: number;
  totalPaymentsAmount: number;
  upcomingMaintenanceCount: number;
  expiringWarrantiesCount: number;
  recentCustomers?: Customer[];
}

export type ListCustomersParams = {
/**
 * 搜尋關鍵字（姓名、電話、地址）
 */
search?: string;
/**
 * 是否包含兩年前的客戶 (true/false)
 */
includeOld?: string;
};

export type ListQuotesParams = {
customerId?: number;
status?: string;
};

export type ListWorkOrdersParams = {
customerId?: number;
status?: string;
};

export type ListPaymentsParams = {
customerId?: number;
};

export type ListWarrantiesParams = {
customerId?: number;
};

export type ListMaintenanceRemindersParams = {
customerId?: number;
status?: string;
/**
 * 是否只顯示即將到期的提醒 (true/false)
 */
upcoming?: string;
};