# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: 晟風工程 ERP API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /customers:
    get:
      operationId: listCustomers
      tags: [customers]
      summary: 列出所有客戶
      parameters:
        - name: search
          in: query
          schema:
            type: string
          description: 搜尋關鍵字（姓名、電話、地址）
        - name: includeOld
          in: query
          schema:
            type: string
          description: 是否包含兩年前的客戶 (true/false)
      responses:
        "200":
          description: 客戶列表
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Customer"
    post:
      operationId: createCustomer
      tags: [customers]
      summary: 新增客戶
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerInput"
      responses:
        "201":
          description: 新增成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
  /customers/{id}:
    get:
      operationId: getCustomer
      tags: [customers]
      summary: 取得客戶詳細資料
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: 客戶資料
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
        "404":
          description: 找不到客戶
    patch:
      operationId: updateCustomer
      tags: [customers]
      summary: 更新客戶資料
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerUpdate"
      responses:
        "200":
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
        "404":
          description: 找不到客戶
    delete:
      operationId: deleteCustomer
      tags: [customers]
      summary: 刪除客戶
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: 刪除成功
        "404":
          description: 找不到客戶
  /customers/{customerId}/ac-units:
    get:
      operationId: listAcUnits
      tags: [acUnits]
      summary: 取得客戶的冷氣設備列表
      parameters:
        - name: customerId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: 冷氣設備列表
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AcUnit"
    post:
      operationId: createAcUnit
      tags: [acUnits]
      summary: 新增冷氣設備
      parameters:
        - name: customerId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AcUnitInput"
      responses:
        "201":
          description: 新增成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AcUnit"
  /ac-units/{id}:
    patch:
      operationId: updateAcUnit
      tags: [acUnits]
      summary: 更新冷氣設備
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AcUnitUpdate"
      responses:
        "200":
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AcUnit"
    delete:
      operationId: deleteAcUnit
      tags: [acUnits]
      summary: 刪除冷氣設備
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: 刪除成功
...
  /maintenance-reminders:
    get:
      operationId: listMaintenanceReminders
      tags: [maintenance]
      summary: 列出保養提醒
      parameters:
        - name: customerId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
        - name: upcoming
          in: query
          schema:
            type: string
          description: 是否只顯示即將到期的提醒 (true/false)
      responses:
        "200":
          description: 保養提醒列表
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MaintenanceReminder"
    post:
      operationId: createMaintenanceReminder
      tags: [maintenance]
      summary: 新增保養提醒
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MaintenanceReminderInput"
      responses:
        "201":
          description: 新增成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MaintenanceReminder"
  /maintenance-reminders/{id}:
    get:
      operationId: getMaintenanceReminder
      tags: [maintenance]
      summary: 取得保養提醒
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: 保養提醒
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MaintenanceReminder"
        "404":
          description: 找不到保養提醒
    patch:
      operationId: updateMaintenanceReminder
      tags: [maintenance]
      summary: 更新保養提醒
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MaintenanceReminderUpdate"
      responses:
        "200":
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MaintenanceReminder"
    delete:
      operationId: deleteMaintenanceReminder
      tags: [maintenance]
      summary: 刪除保養提醒
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: 刪除成功
...
components:
  schemas:
...
    Customer:
      type: object
      required: [id, name, phone, address, createdAt, updatedAt]
      properties:
        id: { type: integer }
        name: { type: string }
        phone: { type: string }
        address: { type: string }
        email: { type: ["string", "null"] }
        discountScheme: { type: ["string", "null"] }
        notes: { type: ["string", "null"] }
        createdAt: { type: string }
        updatedAt: { type: string }
    CustomerInput:
      type: object
      required: [name, phone, address]
      properties:
        name: { type: string, minLength: 1 }
        phone: { type: string, minLength: 1 }
        address: { type: string, minLength: 1 }
        email: { type: string }
        discountScheme: { type: string }
        notes: { type: string }
    CustomerUpdate:
      type: object
      properties:
        name: { type: string, minLength: 1 }
        phone: { type: string }
        address: { type: string }
        email: { type: string }
        discountScheme: { type: string }
        notes: { type: string }
    AcUnit:
      type: object
      required: [id, customerId, brand, model, createdAt, updatedAt]
      properties:
        id: { type: integer }
        customerId: { type: integer }
        brand: { type: string }
        model: { type: string }
        serialNumber: { type: ["string", "null"] }
        purchaseDate: { type: ["string", "null"] }
        installationDate: { type: ["string", "null"] }
        notes: { type: ["string", "null"] }
        createdAt: { type: string }
        updatedAt: { type: string }
    AcUnitInput:
      type: object
      required: [brand, model]
      properties:
        brand: { type: string, minLength: 1 }
        model: { type: string, minLength: 1 }
        serialNumber: { type: string }
        purchaseDate: { type: string }
        installationDate: { type: string }
        notes: { type: string }
    AcUnitUpdate:
      type: object
      properties:
        brand: { type: string }
        model: { type: string }
        serialNumber: { type: string }
        purchaseDate: { type: string }
        installationDate: { type: string }
        notes: { type: string }
...
    MaintenanceReminder:
      type: object
      required: [id, customerId, reminderDate, description, status, createdAt]
      properties:
        id: { type: integer }
        customerId: { type: integer }
        customerName: { type: ["string", "null"] }
        acUnitId: { type: ["integer", "null"] }
        reminderDate: { type: string }
        description: { type: string }
        status: { type: string }
        notes: { type: ["string", "null"] }
        createdAt: { type: string }
    MaintenanceReminderInput:
      type: object
      required: [customerId, reminderDate, description, status]
      properties:
        customerId: { type: integer }
        acUnitId: { type: integer }
        reminderDate: { type: string }
        description: { type: string, minLength: 1 }
        status: { type: string }
        notes: { type: string }
    MaintenanceReminderUpdate:
      type: object
      properties:
        reminderDate: { type: string }
        description: { type: string }
        status: { type: string }
        notes: { type: string }
...