# lib/api-zod/src/generated/types/acUnit.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * 晟風工程 ERP API
 * OpenAPI spec version: 0.1.0
 */

export interface AcUnit {
  id: number;
  customerId: number;
  brand: string;
  model: string;
  /** @nullable */
  serialNumber?: string | null;
  /** @nullable */
  purchaseDate?: string | null;
  /** @nullable */
  installationDate?: string | null;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
  updatedAt: string;
}