# lib/api-zod/src/generated/types/customer.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * 晟風工程 ERP API
 * OpenAPI spec version: 0.1.0
 */

export interface Customer {
  id: number;
  name: string;
  phone: string;
  address: string;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  discountScheme?: string | null;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
  updatedAt: string;
}