# lib/api-zod/src/generated/types/customerUpdate.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * 晟風工程 ERP API
 * OpenAPI spec version: 0.1.0
 */

export interface CustomerUpdate {
  /** @minLength 1 */
  name?: string;
  phone?: string;
  address?: string;
  email?: string;
  discountScheme?: string;
  notes?: string;
}