# lib/api-zod/src/generated/types/maintenanceReminderInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * 晟風工程 ERP API
 * OpenAPI spec version: 0.1.0
 */

export interface MaintenanceReminderInput {
  customerId: number;
  acUnitId?: number;
  reminderDate: string;
  /** @minLength 1 */
  description: string;
  status: string;
  notes?: string;
}